extract useBalances hook
